Enc common: superres
