+++ Source/Lib/Encoder/Codec/EbEntropyCoding.h
@@ -327,8 +327,6 @@ static INLINE int32_t tile_log2(int32_t blk_size, int32_t target) {
     for (k = 0; (blk_size << k) < target; k++) {}
     return k;
 }
-void eb_av1_tile_set_col(TileInfo *tile, const TilesInfo *tiles_info, int32_t mi_cols, int col);
-void eb_av1_tile_set_row(TileInfo *tile, TilesInfo *tiles_info, int32_t mi_rows, int row);
 
 int32_t eb_aom_count_primitive_subexpfin(uint16_t n, uint16_t k, uint16_t v);
 int32_t eb_aom_count_primitive_refsubexpfin(uint16_t n, uint16_t k, uint16_t ref, uint16_t v);
